+++ back/controller/users.go
@@ -1,0 +1,25 @@
+package controller
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+// /users
+func GetUsers(c *gin.Context) {
+	var res User_register
+	if err := c.Bind(&res); err != nil {
+		c.String(http.StatusBadRequest, "bad request")
+		return
+	}
+	c.JSON(http.StatusOK, res)
+}
+func DeleteUsers(c *gin.Context) {
+	var user User_res
+	if err := c.Bind(&user); err != nil {
+		c.String(http.StatusBadRequest, "bad request")
+		return
+	}
+	c.JSON(http.StatusOK, "StatusOK")
+}
